Create project module and add all project logic there
